refactor: SearchResult leave the code small
